Refactoriza vista de login
-Mueve page de login/registro. -refactoriza componente de textfield. -Mueve gradientes a archivo separado
