+++ lib/config/routes/routes.dart
@@ -2,8 +2,8 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../features/authentication/presentation/pages/login/login_page.dart';
-import '../../features/authentication/presentation/pages/register/register_page.dart';
+import '../../features/authentication/presentation/pages/login_page.dart';
+import '../../features/authentication/presentation/pages/register_page.dart';
 import '../../features/stray_dog/presentation/pages/landing/landing.dart';
 
 enum Routes {
