+++ lib/features/authentication/presentation/pages/login_page.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+import '../../../../core/gradients/gradients.dart';
+import '../../../../core/utils/widgets/generic_scaffold.dart';
+import '../bloc/widgets/login_form.dart';
+
+class LoginPage extends StatefulWidget {
+  const LoginPage({super.key});
+
+  @override
+  State<LoginPage> createState() => _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  @override
+  Widget build(BuildContext context) {
+    return GenericScaffold(
+      showAppBar: false,
+      showBottomBar: false,
+      title: "Login_form_title",
+      bodyWidget: Container(
+        decoration: BoxDecoration(
+          gradient: Gradients.loginBackGround(),
+        ),
+        child: LoginForm(),
+      ),
+    );
+  }
+}
