+++ lib/features/authentication/presentation/pages/register_page.dart
@@ -1,11 +1,11 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../bloc/widgets/register_background.dart';
 
-import '../../../../../core/utils/widgets/generic_scaffold.dart';
-import '../../bloc/authentication_bloc.dart';
-import '../../bloc/widgets/register_form.dart';
+import '../../../../core/utils/widgets/generic_scaffold.dart';
+import '../bloc/authentication_bloc.dart';
+import '../bloc/widgets/register_background.dart';
+import '../bloc/widgets/register_form.dart';
 
 class RegisterPage extends StatefulWidget {
   const RegisterPage({super.key});
